% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gupd.R
\name{make_gupd}
\alias{make_gupd}
\title{Generalized Uniform Partition Distribution}
\usage{
make_gupd(prob, n_items, log = FALSE)
}
\arguments{
\item{prob}{An vector of probability weights, where element \eqn{i} is the
unnormalized marginal probabilty of partitions with \eqn{i} clusters.  If
the length of this vector is less than the number of items, zero is implied
for all remaining elements.}

\item{n_items}{An integer for the number of items in the partition.}

\item{log}{Return results on the natural logarithm scale?}
}
\value{
A function to evaluate the Generalized Uniform Partition Distribution.
}
\description{
A function implementing the probability mass function is provided based on
the supplied probabilty vector and number of items.
}
\examples{
f <- make_gupd(c(1, 10, 5), 6)
f(c(1, 1, 2, 2, 3, 3))
f(3)

}
