% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxido.R
\docType{package}
\name{xhp-package}
\alias{xhp}
\alias{xhp-package}
\title{xhp: Exchangeable Hierarchical Partition Distribution}
\description{
One paragraph description of what the package does.
}
\author{
\strong{Maintainer}: First Last \email{name@gmail.com} (\href{https://orcid.org/ORCID-ID}{ORCID})

Other contributors:
\itemize{
  \item Alex Crichton \email{alex@alexcrichton.com} (Rust crate: proc-macro2) [contributor]
  \item David B. Dahl \email{dahl@stat.byu.edu} (Rust crates: roxido, roxido_macro) [contributor]
  \item David Tolnay \email{dtolnay@gmail.com} (Rust crates: proc-macro2, quote, syn, unicode-ident) [contributor]
}

}
\keyword{internal}
